bumped version from 0.1 to 0.2
This update brings significant improvements as well as addressing some of the glaring flaws of 0.1

Although the version number is low (1.0 would indicate I've done everything I can think of to it), this is release is hopefully very stable (let me know if you encounter anything strange), and it should be a lot more fun now that it has interbbs play.

bug fixes : the game exits gracefully and doesn't throw any errors on exit (that was a stopgap solution)

features : BBS to BBS play, list of players to complete trail, 'tombstones', cause of death variable added • implemented a JSON client and DB to record player deaths and those who complete trail. When you enter the game you will see a list of all the people who have beat the trail and what BBS (it's not easy), followed by a list of those who have died and how far they made it before they died and it tells you what they died of and what they wrote on their tombstone. When you die, you will be shown your tombstone and then a flashing red GAME OVER. If you somehow make it to Oregon City (I've only done it once), you'll get your name on the list of people who have made it, which will be ranked by the date people made it. I will host a score server on my BBS, telnet://futureland.grudgemirror.com, by default, which I hope people will use to connect to for aggregation of score data.